refactor(web): third pass - relocates correction types & methods to input-processor
