+++ common/web/input-processor/src/index.ts
@@ -1,5 +1,7 @@
 /// <reference path="../../../predictive-text/src/worker-interface.d.ts" />
 
+export * from './corrections.js';
+export * from './correctionLayout.js';
 export { default as InputProcessor } from './text/inputProcessor.js';
 export { default as ContextWindow } from './text/contextWindow.js';
 export { default as ModelSpec } from './text/prediction/modelSpec.js';
