+++ common/web/input-processor/tsconfig.json
@@ -2,6 +2,8 @@
   "extends": "../tsconfig.kmw-main-base.json",
   "compilerOptions": {
     "baseUrl": "./",
+    // Necessary for ES5 iteration of Map.keys().
+    "downlevelIteration": true,
     "outDir": "build/obj/",
     "tsBuildInfoFile": "build/obj/tsconfig.tsbuildinfo",
     "rootDir": "./src"
